+++ crawler/worker.py
@@ -5,7 +5,6 @@ from utils.download import download
 from utils import get_logger
 import scraper
 import time
-from analysis import analyzer
 
 
 class Worker(Thread):
@@ -19,7 +18,7 @@ class Worker(Thread):
         assert {getsource(scraper).find(req) for req in {"from urllib.request import", "import urllib.request"}} == {-1}, "Do not use urllib.request in scraper.py"
         
         # Set politeness delay from config
-        analyzer.set_politeness_delay(config.time_delay)
+        frontier.set_politeness_delay(config.time_delay)
         
         super().__init__(daemon=True)
         
@@ -30,14 +29,14 @@ class Worker(Thread):
                 self.logger.info("Frontier is empty. Worker stopping.")
                 break
             
-            analyzer.check_domain_politeness(tbd_url)
+            self.frontier.check_domain_politeness(tbd_url)
             
             resp = download(tbd_url, self.config, self.logger)
             
             self.logger.info(
                 f"Downloaded {tbd_url}, status <{resp.status}>, "
                 f"using cache {self.config.cache_server}.")
-            scraped_urls = scraper.scraper(tbd_url, resp)
+            scraped_urls = scraper.scraper(tbd_url, resp, self.frontier)
             for scraped_url in scraped_urls:
                 self.frontier.add_url(scraped_url)
             self.frontier.mark_url_complete(tbd_url)
